# Pairing: remove preCompute and go routines in MillerLoop