@yelhousni 2021-01-11
In the last commit, I applied the same changes to `bls377`.